kickstart: Assume no autologin user by default
Having "meego" as default was causing SDDM to be configured with
autologin even if we didn't want it to.

That was causing SDDM to fail because Maui didn't have that user.

The default made sense a long time ago with MeeGo because looks like
autologin was always enabled when a default session was specified.

This has changed with Maui, one has to explicitely enable that whether
if a default session is specified or not.